+++ packages/nuxt-ripple/mapping/site/index.ts
@@ -2,6 +2,7 @@ import {
   getImageFromField,
   getBody,
   getBodyFromField,
+  getMediaPath,
   getLinkFromField,
   getSiteKeyValues,
   TideSiteApi
@@ -22,9 +23,9 @@ export default {
       if (src.field_site_logo) {
         return {
           href: '/',
-          src: src.field_site_logo.url,
+          src: getMediaPath(src, 'field_site_logo'),
           altText: src.field_site_logo.meta?.alt,
-          printSrc: getImageFromField(src, 'field_print_friendly_logo')?.src
+          printSrc: getMediaPath(src, 'field_print_friendly_logo')
         }
       }