+++ packages/ripple-tide-api/src/utils.ts
@@ -1,1 +1,2 @@
 export { getDpcPkgs } from './utils/index.js'
+export { stripMediaBaseUrl } from './utils/stripMediaBaseUrl.js'