+++ packages/ripple-tide-search/components/global/TideCustomCollection.vue
@@ -286,7 +286,12 @@ onMapResultsHook.value = () => {
   const hookFnName = props.mapConfig?.onResultsHook
   const fns: Record<
     string,
-    (map: any, results: any, locationQuery: any) => Promise<any>
+    (
+      map: any,
+      results: any,
+      locationQuery: any,
+      mapDeadSpace?: any
+    ) => Promise<any>
   > = appConfig?.ripple?.search?.mapResultHooks || {}
 
   if (!hookFnName) {
@@ -301,7 +306,12 @@ onMapResultsHook.value = () => {
     )
   }
 
-  hookFn(rplMapRef.value, mapResults.value, locationOrGeolocation.value)
+  hookFn(
+    rplMapRef.value,
+    mapResults.value,
+    locationOrGeolocation.value,
+    deadSpace.value
+  )
 }
 
 const resultsContainer = computed(
@@ -552,6 +562,7 @@ const locationOrGeolocation = computed(() => {
           v-bind="locationQueryConfig?.props"
           :label="searchListingConfig.labels?.submit"
           :placeholder="searchListingConfig.labels?.placeholder"
+          :variant="reverseFields ? 'reverse' : 'default'"
           :inputValue="locationQuery"
           :resultsloaded="mapFeatures.length > 0"
           :isGettingLocation="isGettingLocation"