+++ packages/ripple-tide-search/components/global/TideSearchListing/TideSearchListingTableUrl.vue
@@ -1,5 +1,6 @@
 <template>
-  <RplTextLink v-if="value" :url="value">{{ value }}</RplTextLink>
+  <RplTextLink v-if="url" :url="url">{{ value }}</RplTextLink>
+  <span v-else>{{ value }}</span>
 </template>
 
 <script setup lang="ts">
@@ -12,8 +13,21 @@ interface Props {
 
 const props = defineProps<Props>()
 
-const value = computed(() => {
+const url = computed(() => {
   const result = getSearchResultValue(props.item, props.column.objectKey, true)
   return Array.isArray(result) ? result[0] : result
+})
+
+const value = computed(() => {
+  if (props.column?.objectTextKey) {
+    const result = getSearchResultValue(
+      props.item,
+      props.column.objectTextKey,
+      true
+    )
+    return Array.isArray(result) ? result[0] : result
+  }
+
+  return url.value
 })
 </script>