+++ packages/ripple-tide-webform/mapping/webforms-mapping.ts
@@ -1,8 +1,21 @@
 import { getBody, TidePageApi } from '@dpc-sdp/ripple-tide-api'
 import { FormKitSchemaNode } from '@formkit/core'
-import { getInputIcons, getValidationAndConditionals } from './webform-utils'
+import {
+  getInputIcons,
+  getValidationAndConditionals
+} from './../server/webform-utils'
 import { getAdvancedAddressMapping } from './webforms-address'
 import type { TideWebformElement, ApiWebForm } from './../types'
+
+interface CustomInputsConfig {
+  [key: string]: {
+    [key: string]: any
+    mapping: Function
+  }
+}
+
+const appConfig = useAppConfig().ripple as CustomInputsConfig
+const customInputs: CustomInputsConfig = appConfig.customInputs || {}
 
 export const getFormSchemaFromMapping = async (
   webform: ApiWebForm,
@@ -319,12 +332,20 @@ export const getFormSchemaFromMapping = async (
         }
         break
       default:
-        mappedField = {
-          $el: 'div',
-          attrs: {
-            class: 'rpl-form__outer rpl-form__input--unsupported'
-          },
-          children: [`"${field['#type']}" is not yet supported`]
+        if (Object.keys(customInputs).includes(field['#type'])) {
+          mappedField = customInputs[field['#type']].mapping(
+            fieldID,
+            field,
+            fieldKey
+          )
+        } else {
+          mappedField = {
+            $el: 'div',
+            attrs: {
+              class: 'rpl-form__outer rpl-form__input--unsupported'
+            },
+            children: [`"${field['#type']}" is not yet supported`]
+          }
         }
     }